react: always extract argTypes, even without parameters.docs
