+++ code/renderers/react/src/entry-preview-docs.ts
@@ -1,29 +1,8 @@
-import type { ArgTypesExtractor } from 'storybook/internal/docs-tools';
-import { enhanceArgTypes, extractComponentDescription } from 'storybook/internal/docs-tools';
-import type { ArgTypesEnhancer, DecoratorFunction } from 'storybook/internal/types';
+import type { DecoratorFunction } from 'storybook/internal/types';
 
-import { extractArgTypes } from './docs/extractArgTypes';
 import { jsxDecorator } from './docs/jsxDecorator';
 import type { ReactRenderer } from './types';
 
-export const parameters: {
-  docs: {
-    story: {
-      inline: boolean;
-    };
-    extractArgTypes: ArgTypesExtractor;
-    extractComponentDescription: (component?: any) => string;
-  };
-} = {
-  docs: {
-    story: { inline: true },
-    extractArgTypes,
-    extractComponentDescription,
-  },
-};
-
 export const decorators: DecoratorFunction<ReactRenderer>[] = [jsxDecorator];
-
-export const argTypesEnhancers: ArgTypesEnhancer<ReactRenderer>[] = [enhanceArgTypes];
 
 export { applyDecorators } from './docs/applyDecorators';
