+++ code/renderers/react/src/extractProps.ts
@@ -7,9 +7,9 @@ import {
 
 import PropTypes from 'prop-types';
 
-import { isMemo } from './lib';
-import { enhancePropTypesProps } from './propTypes/handleProp';
-import { enhanceTypeScriptProps } from './typeScript/handleProp';
+import { isMemo } from './docs/lib';
+import { enhancePropTypesProps } from './docs/propTypes/handleProp';
+import { enhanceTypeScriptProps } from './docs/typeScript/handleProp';
 
 // FIXME
 type Component = any;
